Adding struct for schema and changing event timestamp to int32

+++ pkg/utils/event/event.go
@@ -2,9 +2,9 @@ package event
 
 type Event struct {
 	PartitionKey string         `json:"partitionKey"`
-	Model       string         `json:"model"`
+	Model        string         `json:"model"`
 	EventId      string         `json:"eventId"`
-	Timestamp    int64          `json:"timestamp"`
+	Timestamp    int            `json:"timestamp"`
 	Stage        string         `json:"stage"`
 	Data         map[string]any `json:"data"`
 }
